feat(vs2022_third): C4018 fix signed/unsigned mismatch in NanoSpec — 6 files (batch 2)
Cast int loop variables to size_t/DWORD in comparisons:
- NanoSpec/BCMenu.cpp: 1 instance (GetMenuItemCount→UINT)
- NanoSpec/ChifTransiStress.cpp: 3 instances (i→DWORD)
- NanoSpec/MainFrm.cpp: 2 instances (i→size_t)
- NanoSpec/PointConfirmationDlg.cpp: 1 instance (i→size_t)
- NanoSpec/PointSetupTeaching2PointDlg.cpp: 1 instance (i→size_t)
- NanoSpec/PointSetupTeachingDlg.cpp: 2 instances (i→size_t)

Total: 10 C4018 warnings fixed.
No behavior changes, only explicit casts to resolve type mismatch.

+++ SRC/NanoSpec/ChifTransiStress.cpp
@@ -474,7 +474,7 @@ void CChiefTransiStress::PreMove()
 							break;
 						}
 					}
-					if (l_StressConfig.Line[l_dwLineNo - 1].dwSectionNum <= i) {
+					if (l_StressConfig.Line[l_dwLineNo - 1].dwSectionNum <= (DWORD)i) {
 						TCHAR l_tszLog[128];
 						_stprintf(l_tszLog, _T("No Sections in Line No.%d."), l_dwLineNo);
 						((CChiefView*) m_pcChiefView)->LogChief(l_tszLog);
